+++ utility/doc/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.0.0.42] - 2023-05-31
 
 - Use mockgen and Bus for DI
+- Cross-check the chains handled by a servicer using the Persistence module
 
 ## [0.0.0.41] - 2023-05-25